Splitting consumer from web api - part 1
The consumer is very CPU/Memory hungry and we want to isolate it from the web api.

Also, as we want to avoid downtime during deployments, we want to run
at least 2 instances of authz, which means that we would need to add
some coordination between consumers if we had more than one.

Instead of that, we will just run 2 instances of the web api and one instance
of the consumer.

We expect CI to fail as the full stack test does not start the consumer.